refactor: split large components into focused subcomponents
- ShareModal (416→178 lines): extract LinkTab, QRTab, TabSwitcher,
  useFocusTrap hook, and downloadQRCode utility
- TokenSelect (331→217 lines): extract CustomTokenForm component
- CreateWorkspace (346→330 lines): extract SYNC_STATUS_CONFIG to constants.ts

Improves maintainability without changing public APIs.
