+++ src/features/invoice/ui/TokenSelect.tsx
@@ -11,14 +11,10 @@ import {
 } from '@/shared/ui/select'
 import { NETWORK_TOKENS, type TokenInfo } from '../model/tokens'
 import { cn } from '@/shared/lib/utils'
-import { useTokenMetadata } from '@/entities/token'
-import { Input } from '@/shared/ui/input'
-import { Button } from '@/shared/ui/button'
 import { Badge } from '@/shared/ui/badge'
 import { TokenIcon } from '@/shared/ui/token-icon'
-import { Loader2Icon, SearchIcon, AlertCircleIcon } from '@/shared/ui/icons'
 import { addressesMatch } from '@/shared/lib/validation'
-import { type Address } from 'viem'
+import { CustomTokenForm } from './CustomTokenForm'
 
 /**
  * TokenSelect Component Props
@@ -70,21 +66,13 @@ export function TokenSelect({
   className,
   id,
 }: TokenSelectProps) {
-  // Custom token state
+  // Custom token mode state
   const [isCustomMode, setIsCustomMode] = React.useState(false)
-  const [customAddress, setCustomAddress] = React.useState('')
 
   // Get available tokens for current network
   const availableTokens = React.useMemo(() => {
     return NETWORK_TOKENS[chainId] || []
   }, [chainId])
-
-  // Fetch custom token metadata
-  const {
-    data: customTokenMetadata,
-    isLoading: isLoadingMetadata,
-    isError: hasMetadataError,
-  } = useTokenMetadata(customAddress as Address, chainId)
 
   // Serialize token for Select component (needs string value)
   const serializeToken = React.useCallback((token: TokenInfo | null): string => {
@@ -102,7 +90,9 @@ export function TokenSelect({
 
       try {
         const { symbol, address } = JSON.parse(value)
-        const token = availableTokens.find((t) => t.symbol === symbol && addressesMatch(t.address, address))
+        const token = availableTokens.find(
+          (t) => t.symbol === symbol && addressesMatch(t.address, address)
+        )
         return token || null
       } catch {
         return null
@@ -126,139 +116,35 @@ export function TokenSelect({
     [deserializeToken, onChange]
   )
 
-  const handleCustomSubmit = React.useCallback(() => {
-    if (!customTokenMetadata) return
-
-    onChange({
-      symbol: customTokenMetadata.symbol ?? '',
-      name: customTokenMetadata.name ?? '',
-      address: customAddress,
-      decimals: customTokenMetadata.decimals ?? 18,
-      iconColor: 'bg-zinc-700', // Default color for custom tokens
-      isCustom: true, // Mark as custom/unverified
-    })
-
-    setIsCustomMode(false)
-    setCustomAddress('')
-  }, [customTokenMetadata, customAddress, onChange])
+  const handleCustomSubmit = React.useCallback(
+    (token: TokenInfo) => {
+      onChange(token)
+      setIsCustomMode(false)
+    },
+    [onChange]
+  )
 
   const handleCustomCancel = React.useCallback(() => {
     setIsCustomMode(false)
-    setCustomAddress('')
   }, [])
 
   const selectedValue = serializeToken(value)
 
   // Check if current value is a custom token (case-insensitive address comparison)
   const isCustomToken =
-    value && !availableTokens.find((t) => t.symbol === value.symbol && addressesMatch(t.address, value.address))
+    value &&
+    !availableTokens.find(
+      (t) => t.symbol === value.symbol && addressesMatch(t.address, value.address)
+    )
 
   if (isCustomMode) {
     return (
-      <div
-        className={cn('space-y-3 rounded-lg border border-zinc-800 bg-zinc-950/50 p-4', className)}
-      >
-        <div className="space-y-1.5">
-          <label className="block text-xs font-medium tracking-wide text-zinc-400 uppercase">
-            Contract Address
-          </label>
-          <div className="relative">
-            <Input
-              value={customAddress}
-              onChange={(e) => setCustomAddress(e.target.value)}
-              placeholder="0x..."
-              className="font-mono text-xs"
-            />
-            <div className="pointer-events-none absolute top-1/2 right-3 -translate-y-1/2">
-              {isLoadingMetadata ? (
-                <Loader2Icon className="h-3.5 w-3.5 animate-spin text-violet-500" />
-              ) : (
-                <SearchIcon className="h-3.5 w-3.5 text-zinc-600" />
-              )}
-            </div>
-          </div>
-        </div>
-
-        {/* Loading Skeleton */}
-        {isLoadingMetadata && (
-          <div className="grid grid-cols-2 gap-2">
-            <div className="space-y-1.5">
-              <label className="block text-xs font-medium tracking-wide text-zinc-400 uppercase">
-                Symbol
-              </label>
-              <div className="h-10 animate-pulse rounded-lg border border-zinc-800 bg-zinc-900/50" />
-            </div>
-            <div className="space-y-1.5">
-              <label className="block text-xs font-medium tracking-wide text-zinc-400 uppercase">
-                Decimals
-              </label>
-              <div className="h-10 animate-pulse rounded-lg border border-zinc-800 bg-zinc-900/50" />
-            </div>
-          </div>
-        )}
-
-        {/* Metadata Display */}
-        {!isLoadingMetadata && customTokenMetadata && !hasMetadataError && (
-          <div className="grid grid-cols-2 gap-2">
-            <div className="space-y-1.5">
-              <label className="block text-xs font-medium tracking-wide text-zinc-400 uppercase">
-                Symbol
-              </label>
-              <input
-                className="flex w-full rounded-lg border border-zinc-800 bg-zinc-900/50 px-3 py-2.5 text-sm text-zinc-100 placeholder:text-zinc-600 focus:border-violet-500 focus:ring-1 focus:ring-violet-500/50 focus:outline-none"
-                value={customTokenMetadata.symbol ?? ''}
-                readOnly
-              />
-            </div>
-            <div className="space-y-1.5">
-              <label className="block text-xs font-medium tracking-wide text-zinc-400 uppercase">
-                Decimals
-              </label>
-              <input
-                className="flex w-full rounded-lg border border-zinc-800 bg-zinc-900/50 px-3 py-2.5 text-sm text-zinc-100 placeholder:text-zinc-600 focus:border-violet-500 focus:ring-1 focus:ring-violet-500/50 focus:outline-none"
-                value={customTokenMetadata.decimals ?? ''}
-                readOnly
-              />
-            </div>
-          </div>
-        )}
-
-        {/* Error State */}
-        {!isLoadingMetadata && hasMetadataError && customAddress && (
-          <div className="flex items-start gap-2 rounded-lg border border-red-900/50 bg-red-950/20 p-3">
-            <AlertCircleIcon className="h-4 w-4 shrink-0 text-red-500" />
-            <div className="flex flex-col gap-1">
-              <p className="text-sm font-medium text-red-400">Invalid Token Address</p>
-              <p className="text-xs text-red-500/80">
-                Unable to fetch token metadata. Please verify the address is correct and is an
-                ERC-20 token.
-              </p>
-            </div>
-          </div>
-        )}
-
-        {/* Action Buttons */}
-        <div className="flex gap-2 pt-1">
-          <Button
-            type="button"
-            variant="ghost"
-            size="sm"
-            onClick={handleCustomCancel}
-            className="flex-1 text-xs font-medium text-zinc-400 hover:text-white"
-          >
-            Cancel
-          </Button>
-          <Button
-            type="button"
-            size="sm"
-            onClick={handleCustomSubmit}
-            disabled={!customAddress || !customTokenMetadata || isLoadingMetadata}
-            className="flex-1 bg-violet-600 text-xs font-bold text-white hover:bg-violet-500 disabled:opacity-50"
-          >
-            Add Token
-          </Button>
-        </div>
-      </div>
+      <CustomTokenForm
+        chainId={chainId}
+        onSubmit={handleCustomSubmit}
+        onCancel={handleCustomCancel}
+        className={className}
+      />
     )
   }
 
@@ -277,7 +163,7 @@ export function TokenSelect({
               )}
               <div className="flex flex-col items-start">
                 <div className="flex items-center gap-1.5">
-                  <span className="leading-none font-bold">{value.symbol}</span>
+                  <span className="font-bold leading-none">{value.symbol}</span>
                   {isCustomToken && (
                     <Badge
                       variant="outline"
