+++ src/widgets/share-modal/lib/index.ts
@@ -4,3 +4,5 @@
 
 export * from './types'
 export * from './social-links'
+export * from './use-focus-trap'
+export * from './download-qr'
